refactor(ai): extract remaining residual fat-pin suites under Phase-8 gate
Split develop, economy-filter, priority-weight, conquest-wiring, treasury,
and diplomatic-scoring pin contracts into topical *_cases modules and extend
repo.ai_residual_fat_pin_cases_required to keep the thin contracts enforced.
