+++ tool/check_ai_residual_fat_pin_suite_size.dart
@@ -16,6 +16,12 @@ const Set<String> residualFatPinSuiteGatedBasenames = {
   'observer_goal_phase_test.dart',
   'treasury_planner_treasury_budget_test.dart',
   'economy_planner_regiment_build_input_production_test.dart',
+  'develop_phase_planner_test.dart',
+  'phase_planner_economy_filter_test.dart',
+  'phase_planner_priority_weight_resolvers_test.dart',
+  'phase_planner_conquest_wiring_test.dart',
+  'treasury_planner_test.dart',
+  'diplomatic_candidate_scoring_test.dart',
 };
 
 bool aiResidualFatPinSuiteSizePathInScope(String slashPath) {
